requirements.txt
pybit
binance-connector
pandas
pandas_ta
matplotlib
scipy